+++ resources/views/tailwind/includes/offline.blade.php
@@ -1,5 +1,5 @@
 @if ($offlineIndicator)
-    <div wire:offline>
+    <div wire:offline.class.remove="hidden" class="hidden">
         <div class="rounded-md bg-red-50 p-4 mb-4">
             <div class="flex">
                 <div class="flex-shrink-0">