+++ src/Traits/WithFilters.php
@@ -2,6 +2,10 @@
 
 namespace Rappasoft\LaravelLivewireTables\Traits;
 
+use Illuminate\Database\Eloquent\Builder;
+use Rappasoft\LaravelLivewireTables\Utilities\ColumnUtilities;
+use Rappasoft\LaravelLivewireTables\Views\Column;
+
 /**
  * Trait WithFilters.
  */
@@ -182,5 +186,68 @@ trait WithFilters
     public function getFilterOptions(string $filter): array
     {
         return array_filter(array_keys($this->filters()[$filter]->options() ?? []));
+    }
+
+    /**
+     * Collects columns with $searchable = true
+     *
+     * @return Column[]
+     */
+    public function getSearchableColumns() : array
+    {
+        return array_filter($this->columns(), fn (Column $column) => $column->isSearchable());
+    }
+
+    /**
+     * Apply Search Filter
+     *
+     * @param Builder $query
+     * @return Builder
+     */
+    public function applySearchFilter(Builder $query): Builder
+    {
+        $searchableColumns = $this->getSearchableColumns();
+
+        if ($this->hasFilter('search') && count($searchableColumns)) {
+            $search = trim($this->getFilter('search'));
+
+            // Group search conditions together
+            $query->where(function (Builder $subQuery) use ($search, $query, $searchableColumns) {
+                foreach ($searchableColumns as $column) {
+                    // Does this column have an alias or relation?
+                    $hasRelation = ColumnUtilities::hasRelation($column->column());
+
+                    // Let's try to map this column to a selected column
+                    $selectedColumn = ColumnUtilities::mapToSelected($column->column(), $query);
+
+                    // If the column has a search callback, just use that
+                    if ($column->hasSearchCallback()) {
+                        // Call the callback
+                        ($column->getSearchCallback())($subQuery, $search);
+                    } elseif (! $hasRelation || $selectedColumn) { // If the column isn't a relation or if it was previously selected
+                        $whereColumn = $selectedColumn ?? $column->column();
+
+                        // TODO: Skip Aggregates
+                        if (! $hasRelation) {
+                            $whereColumn = $query->getModel()->getTable() . '.' . $whereColumn;
+                        }
+
+                        // We can use a simple where clause
+                        $subQuery->orWhere($whereColumn, 'like', '%' . $search . '%');
+                    } else {
+                        // Parse the column
+                        $relationName = ColumnUtilities::parseRelation($column->column());
+                        $fieldName = ColumnUtilities::parseField($column->column());
+
+                        // We use whereHas which can work with unselected relations
+                        $subQuery->orWhereHas($relationName, function (Builder $hasQuery) use ($fieldName, $search) {
+                            $hasQuery->where($fieldName, 'like', '%' . $search . '%');
+                        });
+                    }
+                }
+            });
+        }
+
+        return $query;
     }
 }